requirements.txt
autogen-agentchat>=0.4.0
autogen-ext[openai]>=0.4.0
chromadb
openai
python-dotenv
rich
sentence-transformers
tavily-python
